requirements.txt
Flask
flask-cors
pymongo
python-dotenv
PyJWT
bcrypt
cloudinary
